traffic_notifications/requirements.txt
requests
requests_oauthlib
psycopg2-binary
schedule
